#4: Add GraphQL integration to backend and implement ExerciseResolver (#15)
This change introduces GraphQL support to the backend, enabling the use of GraphQL queries and mutations. It also includes an `ExerciseResolver` implementation with sample exercises, facilitating interaction with exercise data through the GraphQL API. This ensures better scalability and a modern API for future client applications.

It uses mock data for now, persistence will be added later.

Reviewed-on: https://git.besi.dev/besi/RepVault/pulls/15
Co-authored-by: Dávid Besenyei <david.besenyei97@gmail.com>
Co-committed-by: Dávid Besenyei <david.besenyei97@gmail.com>