Fix runtime error by adjusting toEntity method usage
The previously used `toEntity` method caused runtime errors, because it tried to set a `null` set group by default, which is not allowed. Thus, the other method is used, where the proper set group is already passed down as a parameter.

The previously used method is now unused. But the removal of it changes the generated code in a way that causes a build issue. Thus, it remains in the codebase.